# Conflicts: #	cmd/main.go #	commands.txt #	db/scripts/apply_operation.sql #	internal/tgbot/account.go #	internal/tgbot/category.go #	internal/tgbot/currency.go #	internal/tgbot/objects/cancelFuncMap.go #	internal/tgbot/payment_operation.go #	internal/tgbot/tgbot.go #	internal/tgbot/transfer_operation.go #	internal/tgbot/user.go #	internal/tgbot/util.go #	pkg/core/exchange/exchange.go #	pkg/core/system_category/system_category.go #	pkg/model/account.go #	pkg/model/account_type.go #	pkg/model/currency.go #	pkg/model/operation.go #	pkg/model/user.go #	pkg/store/repository.go #	pkg/store/sqlstore/accountrepository.go #	pkg/store/sqlstore/accountrepository_test.go #	pkg/store/sqlstore/categoryrepository.go #	pkg/store/sqlstore/categoryrepository_test.go #	pkg/store/sqlstore/operationrepository.go #	pkg/store/sqlstore/operationrepository_test.go #	pkg/store/sqlstore/store.go #	pkg/store/sqlstore/userrepository.go #	pkg/store/sqlstore/userrepository_test.go #	pkg/store/teststore/categoryrepository.go #	pkg/store/teststore/operationrepository.go

+++ pkg/store/teststore/userrepository.go
@@ -14,8 +14,8 @@ type UserRepository struct {
 // Create ...
 func (r *UserRepository) Create(u *model.User) error {
 
-	u.Id = int64(len(r.users) + 1)
-	r.users[int(u.Id)] = u
+	u.Id = len(r.users) + 1
+	r.users[u.Id] = u
 
 	return nil
 }
